refactor: extract shared installAndRunVsixSession from runVsixTests/runVsixUpgradeTest
Both drivers duplicated the same 'download VS Code, --install-extension,
runTests(...)' block (differing only in workspaceDir/extensionTestsPath/env).
Extracted it into installAndRunVsixSession.ts; runVsixTests.ts's single run
and runVsixUpgradeTest.ts's two sequential runs against a shared workspaceDir
now both call it, keeping each script focused on its own vsix-resolution and
workspace/port orchestration.
