+++ tests/upgrade/vsixLifecycle/runVsixTests.ts
@@ -1,13 +1,8 @@
-import {
-  downloadAndUnzipVSCode,
-  resolveCliArgsFromVSCodeExecutablePath,
-  runTests
-} from "@vscode/test-electron";
-import { execFileSync } from "child_process";
 import { mkdtempSync, rmSync } from "fs";
 import { tmpdir } from "os";
 import { join } from "path";
 
+import { installAndRunVsixSession } from "./installAndRunVsixSession";
 import { resolveVsixToTest } from "./resolveVsixToTest";
 import { getFreePorts } from "../utils/freePorts";
 
@@ -20,66 +15,26 @@ import { getFreePorts } from "../utils/freePorts";
  */
 async function main(): Promise<void> {
   const { vsixPath, tempDir: vsixTempDir } = await resolveVsixToTest();
-  const userDataDir = mkdtempSync(join(tmpdir(), "azurite-vsix-userdata-"));
-  const extensionsDir = mkdtempSync(join(tmpdir(), "azurite-vsix-extdir-"));
   const workspaceDir = mkdtempSync(join(tmpdir(), "azurite-vsix-workspace-"));
 
   try {
-    const vscodeExecutablePath = await downloadAndUnzipVSCode();
-    const [cli, ...cliArgs] = resolveCliArgsFromVSCodeExecutablePath(
-      vscodeExecutablePath,
-      // Without this, the helper bakes its own default --user-data-dir/
-      // --extensions-dir (pointing at .vscode-test/) into cliArgs, which
-      // would then be duplicated ahead of - and could shadow - the temp
-      // directories we pass explicitly below.
-      { reuseMachineInstall: true }
-    );
-
-    execFileSync(
-      cli,
-      [
-        ...cliArgs,
-        "--install-extension",
-        vsixPath,
-        "--user-data-dir",
-        userDataDir,
-        "--extensions-dir",
-        extensionsDir
-      ],
-      {
-        stdio: "inherit",
-        // The VS Code CLI can resolve to a .cmd/.bat wrapper on Windows,
-        // which Node refuses to spawn directly without shell: true.
-        shell: process.platform === "win32"
-      }
-    );
-
     // Never probe the well-known default ports here - an already-running
     // developer Azurite instance would be mistaken for this installed VSIX
     // (VSCServerManagerBase.start() swallows EADDRINUSE and the lifecycle
     // suite's probe treats any HTTP response as success).
     const [blobPort, queuePort, tablePort] = await getFreePorts(3);
 
-    await runTests({
-      vscodeExecutablePath,
-      extensionDevelopmentPath: join(__dirname, "driverExtension"),
-      extensionTestsPath: join(__dirname, "suite", "index.js"),
-      extensionTestsEnv: {
+    await installAndRunVsixSession(
+      vsixPath,
+      workspaceDir,
+      join(__dirname, "suite", "index.js"),
+      {
         AZURITE_VSIX_LIFECYCLE_BLOB_PORT: String(blobPort),
         AZURITE_VSIX_LIFECYCLE_QUEUE_PORT: String(queuePort),
         AZURITE_VSIX_LIFECYCLE_TABLE_PORT: String(tablePort)
-      },
-      launchArgs: [
-        workspaceDir,
-        "--user-data-dir",
-        userDataDir,
-        "--extensions-dir",
-        extensionsDir
-      ]
-    });
+      }
+    );
   } finally {
-    rmSync(userDataDir, { recursive: true, force: true });
-    rmSync(extensionsDir, { recursive: true, force: true });
     rmSync(workspaceDir, { recursive: true, force: true });
     if (vsixTempDir) {
       rmSync(vsixTempDir, { recursive: true, force: true });
