+++ tests/upgrade/vsixLifecycle/runVsixUpgradeTest.ts
@@ -1,78 +1,13 @@
-import {
-  downloadAndUnzipVSCode,
-  resolveCliArgsFromVSCodeExecutablePath,
-  runTests
-} from "@vscode/test-electron";
-import { execFileSync } from "child_process";
 import { mkdtempSync, rmSync } from "fs";
 import { tmpdir } from "os";
 import { join } from "path";
 
+import { installAndRunVsixSession } from "./installAndRunVsixSession";
 import {
   packageLocalVsix,
   resolveMarketplaceVsixForUpgrade
 } from "./resolveVsixToTest";
 import { getFreePorts } from "../utils/freePorts";
-
-/**
- * Installs `vsixPath` into a fresh, isolated VS Code user profile and runs
- * the given extensionTestsPath against `workspaceDir` opened as the sole
- * workspace folder (so `azurite.location` defaults to it - see
- * VSCEnvironment.location()). Mirrors runVsixTests.ts's single-session
- * lifecycle, but parameterized so it can be called twice - once per upgrade
- * phase - against the SAME workspaceDir.
- */
-async function runVsixSession(
-  vsixPath: string,
-  workspaceDir: string,
-  extensionTestsPath: string,
-  extensionTestsEnv: NodeJS.ProcessEnv
-): Promise<void> {
-  const userDataDir = mkdtempSync(join(tmpdir(), "azurite-vsix-userdata-"));
-  const extensionsDir = mkdtempSync(join(tmpdir(), "azurite-vsix-extdir-"));
-
-  try {
-    const vscodeExecutablePath = await downloadAndUnzipVSCode();
-    const [cli, ...cliArgs] = resolveCliArgsFromVSCodeExecutablePath(
-      vscodeExecutablePath,
-      { reuseMachineInstall: true }
-    );
-
-    execFileSync(
-      cli,
-      [
-        ...cliArgs,
-        "--install-extension",
-        vsixPath,
-        "--user-data-dir",
-        userDataDir,
-        "--extensions-dir",
-        extensionsDir
-      ],
-      {
-        stdio: "inherit",
-        shell: process.platform === "win32"
-      }
-    );
-
-    await runTests({
-      vscodeExecutablePath,
-      extensionDevelopmentPath: join(__dirname, "driverExtension"),
-      extensionTestsPath,
-      extensionTestsEnv,
-      launchArgs: [
-        workspaceDir,
-        "--user-data-dir",
-        userDataDir,
-        "--extensions-dir",
-        extensionsDir
-      ]
-    });
-  } finally {
-    rmSync(userDataDir, { recursive: true, force: true });
-    rmSync(extensionsDir, { recursive: true, force: true });
-  }
-}
 
 /**
  * Drives the VSIX upgrade compatibility scenario, mirroring what
@@ -108,7 +43,7 @@ async function main(): Promise<void> {
     };
 
     // 1. Seed blob/queue/table data with the OLD (latest published Marketplace) vsix.
-    await runVsixSession(
+    await installAndRunVsixSession(
       oldVsixPath,
       workspaceDir,
       join(__dirname, "upgradeSuite", "seedIndex.js"),
@@ -116,7 +51,7 @@ async function main(): Promise<void> {
     );
 
     // 2. Install the NEW (local build) vsix over the SAME workspace and verify.
-    await runVsixSession(
+    await installAndRunVsixSession(
       newVsixPath,
       workspaceDir,
       join(__dirname, "upgradeSuite", "verifyIndex.js"),
